+++ bot/settings.py
@@ -36,6 +36,12 @@ CPF_AUTH, VOICE_AUTH = range(2)
 # Feedback constants
 FEEDBACK = range(1)
 
+# Auth admin
+EMAIL_AUTH_ADMIN, PASSWORD_AUTH_ADMIN, REPEAT_AUTH_ADMIN = range(3)
+
+# New admin constants
+ADMIN_REGISTER_EMAIL, ADMIN_REGISTER_PWD = range(2)
+
 # Log format
 FORMAT = '%(asctime)s - %(levelname)s - %(name)s - %(message)s'